@public #2948096: Change Text of ButtonControl The following methods have been added to the ButtonControl: setText(), getTextWidth(), getTextHeight(), getFont(), setFont(), setTextVAlign(), setTextHAlign(), setColor(). This methods are for easy access to the underlying <label> elements so now you can change the Buttontext without knowing the internal working of the Control. Just as one would expect.
#2956092: Dynamically Resize Buttons according to the text
The default behaviour of Buttoncontrols is to use the fixed width that isw definied in the style file. This hasn't been changed. If you simply add a button control you will still get a button with a fixed width in pixels that won't change if you change the text. You can specify a different fixed size with the width property to override the default width from the style when you use the button.
What have been changed now is, that when the button has no width set (from either the style or the <control name="button" ...> tag) then the button will always resize to the text content. Because the style always comes with a default width at the moment you'll need to manually override the width when you use the button with an empty string. Example: <control name="button" width="" ..>. This way the button does not have a fixed width anymore and therefore it will be changed to the size of the text. This will even work if you dynamically change the button text from code. Remember that at the moment you still need to manually request a relayout of the screen (screen.layoutLayers()).

Note:
To get the automatic resizing working the CenterLayout layout type has been changed. If you use a childLayout="center" for an element and this element does not have a width constraint and the first child of this element (the actual element that gets centered by the CenterLayout) has a fixed size than the element you attached the "center" layout to it will be resized to the size of the first element. If you don't want that to happen then you always need to specify a width property to the element.