@public Added Attribute "customKey" to <onCustom> effect type. This finally allows multiple onCustom effects on an element that can be triggered from java with the given key. This means that you can add now have multiple <onCustom> effects on an element and only trigger some of them. Before this change is was only possible to trigger them all.
Moved the <onCustom> effect before the <onActive> effect when rendering effects. This might break code that relies on the <onCustom> to be rendered after <onActive> tho. I'm not quite happy with this. A better way would be that the order in which the effects appear in the nifty xml will decide the rendering order too. Tried to get this in but it had too many issues so that must wait for a later version.