@public #3091365: Textfield cursor is shown when the textfield is hidden/shown Problem: The cursor element uses an empty cursor image and the ImageOverlayPulsate effect in onActive() mode. This means the cursor is constantly blinking and needed to be hidden when the textfield did not have the keyboard focus. This worked well as long as all the textfields where never hidden. However, being just another child element of the textfield all show()/hide() calls would inherit to the cursor element. This lead to the display of the blinking cursor even when the textfield didn't had the focus. This especially was problematic when you'd use multiple textfield controls and you would f.i. hide the parent panel. In this case all of the textfields would show the blinking cursor :)
Fix:
The cursor blinking already uses alpha blending to animate the blinking cursor. The ImageOverlayPulsate effect is now being applied to onCustom() instead of onActive(). So basically this allows the textfield control to start and stop the onCustom effect when the control gets/lost the keyboard focus. This means the cursor does not depend on the cursor display state and therefore can be hidden/shown in the normal process of hiding/showing the textfield.