@public #3024993: Change mouse cursor Second part of API changes added. New XML-Tags available:
<registerMouseCursor id="someId" filename="cursor.png" hotspotX="0" hotspotY="0" />
...
<onHover name="changeMouseCursor" id="someId"

The new effect changeMouseCursor can be used to dynamically change mouse cursors. It will save the currently activate one, changes the mouse cursor as long as the effect is active and changes the mouse cursor back to the original one when the effect ends.
You'll need to register mouse cursors first with the registerMouseCursor Tag (or the corresponding NiftyMouse API as previously described).