@public custom Nifty EventBus annotation added: you can now add: @NiftyEventSubscriber(id="nifty-id") or @NiftyEventSubscriber(pattern="nifty-.*") to any method to enable calling that method when the corresponding event for the given element is published by Nifty
Example:
@NiftyEventSubscriber(id="my-id")
public void onSomeNiftyEvent(final String id, SomeNiftyEvent event) {
  ...
}

and Nifty will automatically call that method when SomeNiftyEvent is published for the element with the given id. Please Note: The first parameter of the method has to be a String and the second needs to be the specific Event class you want to listen for.

To enable the Annotations you'll need to call:

NiftyEventAnnoationProcessor.process(this); // to register
NiftyEventAnnoationProcessor.unprocess(this); // to unregister

for the class that contains the event handler methods.