@public when using custom controls it's possible to define in the <controlDefinition> a special element (the childRootId Element) as the target element where all child elements of the <control> will be added.
Example 1:

<controlDefinition name="custom" ... childRootId="childRoot">
  <...>
    <panel id="childRoot">
      <!-- all childs of the <control> will be attached in here -->
    </panel>
  <...>
</controlDefinition>

<control name="custom">
  <label id="label" text="hello world" ...>
</control>

In this case the label with the id="label" will be added as a child to the panel with the id="childRoot" when Nifty resolves controls.

Example 2:

<controlDefinition name="custom" ... childRootId="childRoot">
  <...>
    <panel id="childRoot">
      <image id="image"> <!-- there is already a child element in here -->
    </panel>
  <...>
</controlDefinition>

The childRootId mechanism already existed but in the Example 2 all the child elements of the <control> have been added to the already existing element. This has now been changed to replace. When you would use the control in example 2 without any child elements then the image with id="image" will be kept. If you add child elements to the <control> then the image with id="image" will now be removed.

This change was necessary for the now ListBoxControl which will use its own Label element when you don't define any child elements on the <control>. If you want to use your custom elements for the Listbox, maybe images, you can now add the elements simply as child elements to the <control>. In that case the original Label element will be removed.