@public Added support for "#id" in controls so that you can now use: <controlDefinition name="test"><panel id="#panel"... this works exactly like it does for styles and helps to get global unique ids for the whole screen even when you use the same control multiple times on a screen. the "#<id>" will be replaced by the id of the parent (control) so when you use <control id="someId" name="test"> the panel in the example controlDefinition mentioned above will be set automatically to "test#panel"
Added support for "#id" when finding elements. getElement.findElementWithName("#id") will now find all elements that end with "#id"

Fixed Hover effects deactivate order

Added a new init() method to the Controller interface. The call order is now bind() -> init() -> onStartScreen() and allows to access controls in init() that have been initialized in bind(). Before this change it would be possible to try to access controls in bind() but the accessed control bind() method was not yet called. In this situations you can now access the controls in init(). In init() all bind() methods of all controls have been called.

The ControllerEventListener has now been removed and will be replaced by the EventBus notifications.

Added a check for duplicate Element-Ids at the Screen level. If two elements share the same id you get a warning in the log. This is just a warning because nothing bad will happen at all. But because the EventBus notifications now rely on all elements have an unique id you should make sure that there are no duplicates. The warning should help to discover this cases now.