@public Enable subids (like id="#something") for popups too. When the popup is being created then the actual popup id is being applied to the id (like id="mypopup" for the whole popup will change the sub id of an element to id="mypopup#something" automatically). This is necessary because Nifty 2.0 requires id's to be unique for the whole screen (so that the EventBus mechanism works properly using the id of an element).
Effects are now reset when an element is removed from the screen (I don't know why this was missing? o_O) This is especially required for the new mouse cursor image changing effects. When applied to onHover for instance and the element is being removed the effect is required to be deactivated.