@public maybe =) fixed the Nifty consumption of mouse events problem: this was inspired by mulov again and a bit extended. it works now like this:
- whenever there is a method registered for some mouse event, let's say onClick="someMethod()" and this method is invoked Nifty will now assume that the call will be successful (without knowing that it really is because it is executed as an EndOfFrameAction)
- when an invoke has been successfully scheduled for the onClick then this specific mouse event will not (!) be processed any further. This means if you have another layer below an element with an interact method then this will now never retrieve the event.

I'm not sure if there is a use case for this, where you would like to event to processed by more than one element (when stacked above one another) but if there is, then this will now break and don't work anymore :) But the current changed behaviour would be what you would expect: if one element on top of others already handled the event there is no reason that this event will be processed by other below it (keep in mind that all of the visibleToMouse stuff still applies here).

However there is another aspect to this whole issue:

When Nifty looks for an event handler it does not only take the current controller into account. it will also try to find some other controller in the hierachy up until the screencontroller to send the event too. This process is independent of the processing of the mouse events. Here is a quick example of what I mean. Let's suppose you have this kind of hierachy:

screen controller (A)
  panel without controller
    panel with controller (B)
      control without controller
        control with controller (C)
          panel with controller (D)
            control with controller (E)
            
When Nifty resolves this structure it will know before any event processing that for the Element with controller E there is not one controller it will look for methods but it basically will look for all controllers in its hierachy. So when an event for the element with controller E is about to happen it will look for exampe for the onClick="someMethod()" at the controllers E, D, C, B, A. Would there be a "public void someMethod()" on each of these controllers then all the methods will be called! This process is independ of the mouse event handling and is started when the event is acknowledged and send down for execution. So in this respect you could still have that one event will be called on all controllers kinda thing. If you change your methods so that they would return a boolean you could stop that processing however. Just return true in the someMethod() on controller E and someMethod() will NOT be called on D,C,B or A.

I hope that makes sense at all :)