@public potential fix for restoring active effects when elements are hidden/shown:
when you have active effects which use neverStopRendering=true then these effects basically never stop. when you hide such elements then these effects will be stopped. this was necessary to stop effects that influence global state. there is for instance the changeMouse effect which changes the mouse cursor. imagine what would happen when such effects would not being stopped when an element will be hidden.

on the other hand when you hide/show these elements then all of the neverStopRendering effects would not be restarted automatically which might lead to odd things (one example was the checkbox standard control. it is using a customEffect that gets triggered everytime the checkbox is in checked state. in this case the checkbox mark will be rendered with a neverStopRendering effect (in that case using an imageOverlay effect). what was happening in that case was that hiding and showing the checkbox would make that effect disappear and with it the checkbox mark would disappear too.

this commit will add the following solution:

when an element is hidden Nifty will remember all active effects that have the neverStopRendering flag set to true. when the element is being shown again all of these effects will be automatically started again. this solves this issue for now.
it's not known if other problems will arise with that.