@public #3005778: Reset Nifty and Change RenderDevice runtime changing of the screen resolution is now supported using the new nifty.resolutionChanged() method. the way this works is that you first change the screen resolution for the rendering system of your choice and then call nifty.resolutionChanged() so that Nifty can update all of its internal states that depend on the screen resolution.
please note that it is now required for all RenderDevice implementations to return the "current" screen resolution. implementations should make sure that the get Width() and getHeight() calls don't cache the resolution of the initial display but always need to return the current screen resolution.
@private
this change triggered quite a bit of internal changes. especially controls are now notified when a layout of the current screen is happening so that controls are able to update their internal states to the new screen resolution.