fix show tab problem on initial show
the problem: if the tabcontrol is hidden (or is a part of another panel
that is hidden) then the individual tabs will be set to visible when the
panel is being shown (this means all child elements will be shown as
well, which includes the tab). we now kinda work around this by
listening for the EventShowEvent and displaying the correct tab while
explicitly hiding all the others.