Added getWidth(String text, float size) method to RenderFont Interface
This will now allow users to request the width of a text string when the
given size parameter is applied. This is necessary to allow some effects
(f.i. TextSize) to calculate the correct offset of a text when being
rendered with a size different from 1.0.

PS: I hate changing an interface in the 1.3.1 branch (again) but this
one is necessary to fix the TextSize effect that the autoscale feature
broke :/