Fixed the collapse/expand issue. Turns out repeated calls to element.findNiftyControl(name, ListBox.class); give back an empty listbox each time. Storing the initially returned one in a class variable solved that.
Also remove the resetting of the selected item in favor of setting a boolean
on the TreeEntryModelClass.
This gives the ability to add a custom style for the active tree entry later.

Lasty I removed the listBox.clear in favor of a removeAbsolete, which removes
all items in the listbox which are not visible in the tree according to the
expanded status of the tree nodes. The adding now happens based on the index
parameter in the listbox item list.