Consolidate loading of resources and made it the configuration none static
The process of loading resources in all Nifty projects has now be
consolidated using the NiftyResourceLoader class. This especially made
it possible to make the configuration of ResourceLocations none
static. This now allows each Nifty instance to be configured
individually. Attention: To give RenderDevice, SoundDevice and
InputSystem implementations access to the NiftyResourceLoader a new
method in all three interfaces has been introduced: void
setResourceLoader(NiftyResourceLoader loader) that must now be
implemented as well.