default align and valign of childlayout center childs will be center
I might regret that later but when when the parent of an element has 
childLayout="center" we should default the align and valign
values to center instead of the old defaults left and top. I'm
currently not sure if this breaks something somewhere where
where somebody relies on the old defaults. But the new defaults really
make much more sense! :)