align and valign defaults are now defined inside of the layout managers
before the element tried to decide the defaults which didn't work in
all cases, especially when dynamically modifing the parent
relationship. now a new verticalDefault and horizontalDefault enum
value exists that is used when no other value for the alignment is
set. the final alignment value is now being decided inside of the
indidivual LayoutManagers (where this belongs in the first place!).
kudos to mkaring for the hint on how to solve that finally! :)