#3520041: Window DragNDrop multiplies events
In general when a popup is being shown (nifty.showPopup()) all of the
controls it contains are initialized (and in case of the button
control are registered with the EventBus). When the popup is closed
all controls are automatically removed from the EventBus. So popups work
in the general case correctly.

But when a drag operation starts the content of the draggable was being
moved to a newly created popup (with a call to markForMove()) and after
that the popup was displayed (using nifty.showPopup()). So in that case
all the controls - the content of the draggable that have been
initialized already - would be registered AGAIN when the content moves
to the popup! And because when the drag operation ends the content is
moved back and then the popup is closed - and since the content is
already gone the popup when being closed doesnt contained any elements
anymore. So the removal from the EventBus in the button case didn't work
at all too :)

So after the long explanation the fix actual was easy: when the drag
operation starts we will now show the popup first and THEN move the
content to the popup! This way no duplicate registering with the
EventBus is happening and everything should work as expected again!

PS: Another bug has been fixed as well: when the drag operation ends
Nifty looks for droppables at the current position. While doing this
Nifty didn't check if the droppables are visible! So it was possible
that you have some hidden droppable and when you incidentally drop
something at that position it would be possible to drop a draggable to a
hidden droppable - with the effect of your droppably disappearing! Crap!
But thats fixed now! :)