fixed auto-resize-if-all-childs-have-fixed-px-size mode
When all child elements of a parent element have a fixed width or height
value and the parent does not have any width or height constraints set
then Nifty will automatically resize the parent to the sum or the max
width or height value of all child elements. The specifiy Layoutmanager
(vertical, horizontal or center) decides if sum or max is used when
calculating the parent size.

Additionally another problem has been fixed:
when the auto-resize-if-all-childs-have-fixed-px-size mode has
calculated a new constraint for the parent and later the child elements
change in a way that doesn't allow Nifty to calculate a constraint (like
one child is removed or added or a constraint is changed from px to
percent) then the calculated width/height constraint for the parent will
be removed.