new feature: element attribute "renderOrder" to modify the render order
Elements will be rendered ascending by the renderOrder value which must
be an int value. The default value for "renderOrder" is 0 which means
that the original render order of the elements will be used. The render
order is relevant for the child elements of the same parent.

Internally the index in the elements list is used as the
default renderOrder value when you don't set a value. So when you have a
panel with 3 child elements and you don't set a specific renderOrder
value then internally they will automatically get the values 0, 1 and 2.
Rendering occurs in this order then.

You can change the renderOrder value to some high value, like 10000 to
enforce this element to be rendered above all of the other elements or
some very small number like -10000 to render it below all others. You
can even take control of the render order for all elements if you
provide a renderOrder value for all child elements.

Please note: When you use the same renderOrder for two elements then the
id will be used for comparison. If the elements don't have ids or the
same id then object.toString() is being compared (the reference in
memory).