Some warning elimination and code simplification.
Warning elimination:

Removed unused imports.

Commented out unused private members and locals.
I was unsure about SpecialValuesReplaceTest.MyObjectCallback.getValue, so I
just left it in and added a @SuppressWarning.

Changed EffectsType.translateSpecialValues so that it uses Arrays.asList
instead of a generic varargs parameter (which is a combination that the Java
compiler cannot properly check for type safety).

Code simplification:

Removed gratuitous parameter on private function
PaulsSoundSystemSoundDevice.addAdditionalCodecs.

Removed the unused generic type parameter on NiftyEvent and its subclasses.
This is a compatible change, except for applications that try to Do the Right
Thing and added a <?> parameter (which can simply be removed).
Note that generic implementors of NiftyEvent have stayed generic. See
MenuItemActivatedEvent for an example.