#3489461: Hint effect requires interact tag
Automatically enable visibleToMouse for elements that have a hover
effect attached to. this is similar to what we do for the <interact
onClick="" /> way to register mouse click handlers. If you attach
<interact onClick="" /> to an element we automatically enable
visibleToMouse for the element too. I think it makes sense to do the
same for the hover effect (since it doesn't make much sense to add hover
effects to an element and don't enable visibleToMouse for the element)