replaced screen.layoutLayers() with getParent().layoutElements()
UpdateCursor called screen.layoutLayers() when the position of the
cursor needs to be changed. This looks a bit silly (trigger a complete
relayout of the whole screen?) because it performs too much work. On the
other hand it didn't work correctly in all cases when textfields have
been used in popups, for example in the controls demo it's possible to
display a console window as a popup while pressing F1 at any time. The
very first time the console popup was displayed the cursor was rendered
in the upper left corner of the screen. This was because
screen.layoutLayers() does not take the popups into account!

This commit fixed this issue. There still is a bit uncertainty around
this change since the original code was probably there for a reason :)
But so far I didn't notice any problems with it.