The problem is the way Nifty propagates some effects to child elements.
When an effect starts for an element the effect will automatically
forwarded to all child elements. So for an onClick on a panel with child
elements this means that the onClick effect will be started for the
child elements as well. So this might sometimes be used to trigger
effects for child elements when the parent has been clicked.

Since I've not found any control where this feature has actually been
used I've now changed this behaviour. Now onClick will not travel to
child elements anymore. This solves this issue. There is a slight
possibility that this change will break some existing code. So I was for
a long time unsure to change this.

What convinced be to change this in the end were two things:

1. the issue described in here really sucks bad when it occurs :)
2. the same policy to not forward effect activation is already used
successfully for onHover, onStartHover, onEndHover, onEnabled,
onDisabled