#56 TextRender does not wrap new strings
Changing the text with word wrapping enabled needs to trigger a relayout
of the actual text element. Unfortunately the Element the TextRenderer
is attached to is not available in the TextRenderer class. However
changes to the text can actually trigger changes to the size of the
text element. So there is a relayout necessary.

To free users of calling element.layoutElements() by themself this is
now performed in the setText() method. To do this the element the
TextRenderer is attached to is stored in the TextRenderer when the
element has been layouted at least once.