fixed DraggableControl not processing mouse events in some cases
It was possible for DraggableControls to return "processed = false"
for mouse events when it was being dragged very fast or when the mouse
leaves the screen while it was dragged. This should be fixed now.

Implementation note:
Nifty already supported exclusive mouse events for elements that are
focusable. When such an element was initially clicked the element
internally was marked as the one element that captured the mouse. Nifty
made sure that all mouse events were only sent to this element.

However it still was possible to have mouse events that didn't hit any
element at all and therefore were returned as "processed = false". This
has now been changed. Now all mouse events will be send to the element
that has captured the mouse no matter if the mouse hovers the element or
not.

Additionally when a DraggableControl is being dragged it is
now automatically marked as exclusive.