should fixed issue #52 Listbox Focus Problem
The problem was that the Listbox of the DropDown is hidden but it's
still being updated with all the data, selection and the correct focus
item. The problem is that without showing the Listbox at least once
Nifty does not know the layout of the Listbox and the Scrollbar. So
internally the handle position is kept at 0. This has now being fixed
with an explicit (internal) update of the position when the Listbox (and
the scrollbar) is being shown initially.