onInitialClick() will now ALWAYS capture the mouse
This is done to prevent losing mouse events when you click and hold the
mouse button on some GUI element and then move the mouse away from the
control. This kind of capturing the mouse already worked well for all
controls that are focusable. Some controls - like the buttons of
a Scrollbar control - are not supposed to be focusable individually. For
these controls the mouse was not captured but it is now.