CheckBox: Prevent toggle animation when there is no change in state
When you uncheck a CheckBox that is already unchecked or you check a
CheckBox that is already checked then now there is nothing changed. This
prevents the somewhat ugly display of a check mark that disappears when
a CheckBox is initially unchecked and is set explicitly to unchecked by
code again.