ignoreMouseEvents and ignoreKeyboardEvents added at element level
The new methods to ignore mouse events and keyboard events at the global
Nifty level have now been added at the element level as well. You can
now ignore mouse and keyboard events for individual elements too! Please
note that this is a bit redundant to the existing visibleToMouseEvents
but we've kept the same names to be consistent with the method on the
main Nifty instance.

We see it so: visibleToMouseEvents should be used to generally define
visibility to mouse events and ignoreMouseEvents / ignoreKeyboardEvents
to temporarily disable access.