reactivate elements in resetLayers()
The onEndScreen effect will disable interaction with elements by setting
a done boolean to true. Unfortunately this flag was never reset when you
gotoScreen() to the same screen. It was therefore possible for elements
to be visible but not being able to interact with them. Now done is
explicitly reset to false in the startScreen() method.