mouse events have been returned as processed = false in some wrong cases
It was possible that Nifty returned processed = false for events that
from a intuitive standpoint should have been processed. Example:
Clicking on a DropDownControl so that it opens and while it is
opening releasing the mouse button. This was reported as processed =
false which would make it possible for your game to handle the event -
although that would be very odd in that special case.

The problem is that Nifty will block interaction while
onStartScreen/onEndScreen/interactionBlocked-Flag is active. This is
expected in most cases but since these elements could not be interacted
with Nifty return processed = false. This has now been corrected:
Elements that are generally available for interaction but are not at the
moment will still be treated as processed = true.