ListBox improvement: add/insert/remove operation keep displayed elements
This is done by getting the first currently visible item prior of
modification the list of elements and restoring the visible item after
the operation. This seems to work quite well and removes the otherwise
odd possibility of elements being pushed out of view. Kudos to fbucur
and zarch for this suggestion!

The only known side effect is that the update of the visible elements
can now happen twice. With more work this could be prevented at the
expense of mixing otherwise unrelated operations
(updateViewTotalCount() and restoreVisibleItem(visibleItem)). However
since the duplicated update of the labels only happens once for a single
add/remove/addAll/removeAll call this shouldn't really be that bad.